LPD-46909 Extract disableFormSubmitButton function to be shared by the edit JS client extensions page.
